answers

+++ src/main/java/com/example/demo/repositories/UserRepository.java
@@ -1,6 +1,7 @@
 package com.example.demo.repositories;
 
 import com.example.demo.models.User;
+import io.micrometer.common.lang.Nullable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +11,7 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
 
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
     Optional<User> findByEmail(String email);
     Optional<User> findByUsername(String username);
